% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.sace}
\alias{print.sace}
\title{Print results of \code{sace}}
\usage{
\method{print}{sace}(x, ...)
}
\arguments{
\item{x}{an object of class \code{sace}.}

\item{...}{additional arguments.}
}
\value{
the input object is returned silently.
}
\description{
\code{print.sace} prints estimation of the SACE (survivor average causal effect).
}

