% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.sace}
\alias{summary.sace}
\title{Summarize results of \code{sace}}
\usage{
\method{summary}{sace}(object, ...)
}
\arguments{
\item{object}{an object of class \code{sace}.}

\item{...}{additional arguments.}
}
\value{
the input object is returned silently.
}
\description{
\code{summary.sace} summary estimation of the SACE (survivor average causal effect) and all other model parameters.
}
\note{
If \code{need.variance} is \code{TRUE}, \code{sace} must have been called with \code{need.variance == TRUE}, so that the information needed was recorded.
}

